Criação de uma classe para formatar a moeda

+++ src/main/java/com/api/barbershop/utils/FormatCurrency.java
@@ -1,5 +1,6 @@
 package com.api.barbershop.utils;
 
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.util.Locale;
 
@@ -7,7 +8,7 @@ public class FormatCurrency {
 
     private static final Locale PT_BR = new Locale("pt", "BR");
 
-    public static String format(Double value) {
+    public static String format(BigDecimal value) {
         if (value == null) {
             return "R$ 0,00";
         }
